api/requirements.txt
fastapi
uvicorn
motor
redis
python-socketio
pydantic
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.0,<5
